Rename reviews_submit command to send_new_reviews_to_choices_api
